Move version provider out of util too.
