+++ data/beaconrestapi/build.gradle
@@ -7,6 +7,7 @@ dependencies {
     implementation project(':infrastructure:async')
     implementation project(':infrastructure:http')
     implementation project(':infrastructure:logging')
+    implementation project(':infrastructure:version')
     implementation project(':networking:eth2')
     implementation project(':services:serviceutils')
     implementation project(':storage')
