+++ data/beaconrestapi/src/main/java/tech/pegasys/teku/beaconrestapi/handlers/v1/node/GetVersion.java
@@ -28,7 +28,7 @@ import org.jetbrains.annotations.NotNull;
 import tech.pegasys.teku.api.response.v1.node.Version;
 import tech.pegasys.teku.api.response.v1.node.VersionResponse;
 import tech.pegasys.teku.provider.JsonProvider;
-import tech.pegasys.teku.util.cli.VersionProvider;
+import tech.pegasys.teku.infrastructure.version.VersionProvider;
 
 public class GetVersion implements Handler {
   private final JsonProvider jsonProvider;
