+++ infrastructure/version/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+  implementation 'org.apache.commons:commons-lang3'
+  implementation 'org.apache.tuweni:tuweni-bytes'
+}
