+++ networking/p2p/build.gradle
@@ -7,6 +7,7 @@ dependencies {
   implementation project(':infrastructure:logging')
   implementation project(':infrastructure:metrics')
   implementation project(':infrastructure:subscribers')
+  implementation project(':infrastructure:version')
   implementation project(':pow')
   implementation project(':services:serviceutils')
   implementation project(':storage')
