+++ services/beaconchain/build.gradle
@@ -11,9 +11,10 @@ dependencies {
   implementation project(':ethereum:datastructures')
   implementation project(':ethereum:weaksubjectivity')
   implementation project(':infrastructure:async')
+  implementation project(':infrastructure:logging')
+  implementation project(':infrastructure:version')
   implementation project(':networking:p2p')
   implementation project(':networking:eth2')
-  implementation project(':infrastructure:logging')
   implementation project(':pow')
   implementation project(':protoarray')
   implementation project(':services:serviceutils')
