+++ settings.gradle
@@ -29,6 +29,7 @@ include 'infrastructure:unsigned'
 include 'infrastructure:logging'
 include 'infrastructure:subscribers'
 include 'infrastructure:time'
+include 'infrastructure:version'
 include 'networking:eth2'
 include 'networking:nat'
 include 'networking:p2p'
