+++ teku/build.gradle
@@ -16,6 +16,7 @@ dependencies {
   implementation project(':infrastructure:events')
   implementation project(':infrastructure:async')
   implementation project(':infrastructure:logging')
+  implementation project(':infrastructure:version')
   implementation project(':networking:p2p')
   implementation project(':networking:eth2')
   implementation project(':networking:nat')
