+++ teku/src/main/java/tech/pegasys/teku/ValidatorNode.java
@@ -34,7 +34,7 @@ import tech.pegasys.teku.infrastructure.time.SystemTimeProvider;
 import tech.pegasys.teku.service.serviceutils.ServiceConfig;
 import tech.pegasys.teku.service.serviceutils.layout.DataDirLayout;
 import tech.pegasys.teku.services.ValidatorNodeServiceController;
-import tech.pegasys.teku.util.cli.VersionProvider;
+import tech.pegasys.teku.infrastructure.version.VersionProvider;
 import tech.pegasys.teku.util.config.Constants;
 
 public class ValidatorNode implements Node {
