+++ teku/src/test/java/tech/pegasys/teku/cli/options/LoggingOptionsTest.java
@@ -20,7 +20,7 @@ import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.cli.AbstractBeaconNodeCommandTest;
 import tech.pegasys.teku.infrastructure.logging.LoggingConfig;
 import tech.pegasys.teku.infrastructure.logging.LoggingDestination;
-import tech.pegasys.teku.util.cli.VersionProvider;
+import tech.pegasys.teku.infrastructure.version.VersionProvider;
 
 public class LoggingOptionsTest extends AbstractBeaconNodeCommandTest {
